requirements.txt
requests
typing
python-dotenv